Moved create_and_assign_tags to helpers file
