+++ backend/tests/functional_tests/conftest.py
@@ -11,9 +11,7 @@ from medtagger.database.fixtures import apply_all_fixtures
 from medtagger.clients.hbase_client import HBaseClient
 from medtagger.api.auth.business import create_user, sign_in_user
 from medtagger.api.users.business import set_user_role
-from medtagger.repositories.label_tag import LabelTagRepository
 from medtagger.repositories.roles import RolesRepository
-from medtagger.repositories.scan_categories import ScanCategoriesRepository
 
 logger = logging.getLogger(__name__)
 
@@ -66,16 +64,3 @@ def get_token_for_logged_in_user(role: str) -> str:
 
     with app.app_context():
         return sign_in_user(email, password)
-
-
-def create_tag_and_assign_to_category(key: str, name: str, scan_category_key: str) -> str:
-    """Create Label Tag and assign it to Scan Category.
-
-    :param key: key that will identify such Label Tag
-    :param name: name that will be used in the User Interface for such Label Tag
-    :param scan_category_key: key of the Scan Category that Label Tag will be assigned to
-    :return: Label Tag key
-    """
-    label_tag = LabelTagRepository.add_new_tag(key, name)
-    ScanCategoriesRepository.assign_label_tag(label_tag, scan_category_key)
-    return label_tag.key
